+++ poem/src/listener/mod.rs
@@ -1,5 +1,8 @@
 //! Commonly used listeners.
 
+#[cfg(feature = "acme")]
+#[cfg_attr(docsrs, doc(cfg(feature = "acme")))]
+pub mod acme;
 mod combined;
 #[cfg(any(feature = "native-tls", feature = "rustls"))]
 mod handshake_stream;
@@ -20,9 +23,12 @@ use std::{
     task::{Context, Poll},
 };
 
+use futures_util::{future::BoxFuture, FutureExt, TryFutureExt};
 use http::uri::Scheme;
 use tokio::io::{AsyncRead, AsyncWrite, ReadBuf, Result as IoResult};
 
+#[cfg(feature = "acme")]
+use self::acme::{AutoCert, AutoCertListener};
 #[cfg(any(feature = "native-tls", feature = "rustls"))]
 pub use self::handshake_stream::HandshakeStream;
 #[cfg(feature = "native-tls")]
@@ -71,7 +77,7 @@ pub trait AcceptorExt: Acceptor {
         Combined::new(self, other)
     }
 
-    /// Wrap the acceptor in a Box.
+    /// Wrap the acceptor in a `Box`.
     fn boxed(self) -> BoxAcceptor
     where
         Self: Sized + 'static,
@@ -159,6 +165,42 @@ pub trait Listener: Send {
     {
         NativeTlsListener::new(self, config_stream)
     }
+
+    /// Consume this listener and return a new ACME listener.
+    ///
+    /// # Example
+    ///
+    /// ```
+    /// use poem::listener::{
+    ///     acme::{AutoCert, LETS_ENCRYPT_PRODUCTION},
+    ///     Listener, TcpListener,
+    /// };
+    ///
+    /// let listener = TcpListener::bind("0.0.0.0:443").acme(
+    ///     AutoCert::builder()
+    ///         .directory_url(LETS_ENCRYPT_PRODUCTION)
+    ///         .domain("poem.rs")
+    ///         .build()
+    ///         .unwrap(),
+    /// );
+    /// ```
+    #[cfg(feature = "acme")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "acme")))]
+    #[must_use]
+    fn acme(self, auto_cert: AutoCert) -> AutoCertListener<Self>
+    where
+        Self: Sized,
+    {
+        AutoCertListener::new(self, auto_cert)
+    }
+
+    /// Wrap the listener in a `Box`.
+    fn boxed(self) -> BoxListener
+    where
+        Self: Sized + 'static,
+    {
+        BoxListener::new(self)
+    }
 }
 
 #[async_trait::async_trait]
@@ -167,6 +209,15 @@ impl Listener for Infallible {
 
     async fn into_acceptor(self) -> IoResult<Self::Acceptor> {
         unreachable!()
+    }
+}
+
+#[async_trait::async_trait]
+impl<T: Listener + Sized> Listener for Box<T> {
+    type Acceptor = T::Acceptor;
+
+    async fn into_acceptor(self) -> IoResult<Self::Acceptor> {
+        (*self).into_acceptor().await
     }
 }
 
@@ -241,6 +292,25 @@ impl AsyncWrite for BoxIo {
     fn poll_shutdown(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Result<(), Error>> {
         let this = &mut *self;
         Pin::new(&mut this.writer).poll_shutdown(cx)
+    }
+}
+
+/// An owned dynamically typed Listener for use in cases where you can’t
+/// statically type your result or need to add some indirection.
+pub struct BoxListener(BoxFuture<'static, IoResult<BoxAcceptor>>);
+
+impl BoxListener {
+    fn new<T: Listener + 'static>(listener: T) -> Self {
+        BoxListener(listener.into_acceptor().map_ok(AcceptorExt::boxed).boxed())
+    }
+}
+
+#[async_trait::async_trait]
+impl Listener for BoxListener {
+    type Acceptor = BoxAcceptor;
+
+    async fn into_acceptor(self) -> IoResult<Self::Acceptor> {
+        self.0.await
     }
 }